@@ -3,12 +3,12 @@ import { InternalResponseFormValues, ResponseFormComponent, ResponseFormValues }
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { SharedModule } from '../../../shared/shared.module';
 import { LoggerTestingModule } from 'ngx-logger/testing';
-import { JsonEditorComponent } from 'ang-jsoneditor';
 import { MatCardModule } from '@angular/material/card';
 import { StatusCodes } from 'http-status-codes';
 import { ResponseType } from 'src/app/models/mock-definition/scenario/response-type';
 import { AbstractControl, FormControl } from '@angular/forms';
 import { SimpleChanges } from '@angular/core';
+import { JsonEditorComponent } from 'ang-jsoneditor';
 
 describe('ResponseFormComponent', () => {
   let component: ResponseFormComponent;
